fix spawn food in snake (I hope), codestyle

+++ include/view.hpp
@@ -1,13 +1,13 @@
 #ifndef VIEW_HPP
 #define VIEW_HPP
 
+#include <chrono>
 #include <functional>
 #include <iostream>
 #include <list>
 #include <memory>
 #include <string_view>
 #include <unordered_map>
-#include <chrono>
 
 #include "model.hpp"
 
@@ -25,13 +25,13 @@ namespace graphicInterface {
 
         virtual void run () = 0;
 
-        using coord_t = std::pair<unsigned short, unsigned short>;
+        using coord_t = std::pair<int, int>;
 
         virtual coord_t getTermSize () const = 0;
-        virtual void paint (const coord_t &rabbit) = 0;                         // rabbit
-        virtual void paint (const Control::Snake &snake) = 0;                   // snake
-        virtual void drawFrame () = 0;                                          // frame
-        virtual void write (const std::pair <std::string, size_t> &line) = 0;   // text
+        virtual void paint (const coord_t &rabbit) = 0;                       // rabbit
+        virtual void paint (const Control::Snake &snake) = 0;                 // snake
+        virtual void drawFrame () = 0;                                        // frame
+        virtual void write (const std::pair<std::string, size_t> &line) = 0;  // text
         virtual void endHandler () = 0;
 
         std::function<void ()> drawing;
